+++ RT/code/include/image.hpp
@@ -23,7 +23,7 @@ void WriteByte( FILE* file, unsigned char b )
     assert( success == 1 );
 }
 
-unsigned char ClampColorComponent( float c )
+unsigned char ClampColorComponent( double c )
 {
     int tmp = int( c * 255 );
     
@@ -48,7 +48,7 @@ public:
     Image(int w, int h) {
         width = w;
         height = h;
-        data = new Vector3f[width * height];
+        data = new Vector3d[width * height];
     }
 
     ~Image() {
@@ -63,19 +63,19 @@ public:
         return height;
     }
 
-    const Vector3f &GetPixel(int x, int y) const {
+    const Vector3d &GetPixel(int x, int y) const {
         assert(x >= 0 && x < width);
         assert(y >= 0 && y < height);
         return data[y * width + x];
     }
 
-    void SetAllPixels(const Vector3f &color) {
+    void SetAllPixels(const Vector3d &color) {
         for (int i = 0; i < width * height; ++i) {
             data[i] = color;
         }
     }
 
-    void SetPixel(int x, int y, const Vector3f &color) {
+    void SetPixel(int x, int y, const Vector3d &color) {
         assert(x >= 0 && x < width);
         assert(y >= 0 && y < height);
         data[y * width + x] = color;
@@ -97,7 +97,7 @@ private:
 
     int width;
     int height;
-    Vector3f *data;
+    Vector3d *data;
 
 };
 
@@ -130,7 +130,7 @@ void Image::SaveTGA( const char* filename) const
     {
         for (int x = 0; x < width; x++)
         {
-            Vector3f v = GetPixel(x,y);
+            Vector3d v = GetPixel(x,y);
             // note reversed order: b, g, r
             WriteByte(file,ClampColorComponent(v[2]));
             WriteByte(file,ClampColorComponent(v[1]));
@@ -171,7 +171,7 @@ Image* Image::LoadTGA(const char *filename) {
             b = ReadByte(file);
             g = ReadByte(file);
             r = ReadByte(file);
-            Vector3f color(r/255.0,g/255.0,b/255.0);
+            Vector3d color(r/255.0,g/255.0,b/255.0);
             answer->SetPixel(x,y,color);
         }
     }
@@ -198,7 +198,7 @@ void Image::SavePPM(const char *filename) const {
     // flip y so that (0,0) is bottom left corner
     for (int y = height-1; y >= 0; y--) {
         for (int x=0; x<width; x++) {
-            Vector3f v = GetPixel(x,y);
+            Vector3d v = GetPixel(x,y);
             fputc (ClampColorComponent(v[0]),file);
             fputc (ClampColorComponent(v[1]),file);
             fputc (ClampColorComponent(v[2]),file);
@@ -234,7 +234,7 @@ Image* Image::LoadPPM(const char *filename) {
             r = fgetc(file);
             g = fgetc(file);
             b = fgetc(file);
-            Vector3f color(r/255.0,g/255.0,b/255.0);
+            Vector3d color(r/255.0,g/255.0,b/255.0);
             answer->SetPixel(x,y,color);
         }
     }
@@ -287,7 +287,7 @@ Image::SaveBMP(const char *filename)
     int i, j, ipos;
     int bytesPerLine;
     unsigned char *line;
-    Vector3f*rgb = data;
+    Vector3d*rgb = data;
     FILE *file;
     struct BMPHeader bmph;
