+++ RT/code/include/ray.hpp
@@ -3,7 +3,7 @@
 
 #include <cassert>
 #include <iostream>
-#include <Vector3f.h>
+#include <Vector3d.h>
 
 
 // Ray class mostly copied from Peter Shirley and Keith Morley
@@ -11,7 +11,7 @@ class Ray {
 public:
 
     Ray() = delete;
-    Ray(const Vector3f &orig, const Vector3f &dir) {
+    Ray(const Vector3d &orig, const Vector3d &dir) {
         origin = orig;
         direction = dir;
     }
@@ -21,22 +21,22 @@ public:
         direction = r.direction;
     }
 
-    const Vector3f &getOrigin() const {
+    const Vector3d &getOrigin() const {
         return origin;
     }
 
-    const Vector3f &getDirection() const {
+    const Vector3d &getDirection() const {
         return direction;
     }
 
-    Vector3f pointAtParameter(float t) const {
+    Vector3d pointAtParameter(double t) const {
         return origin + direction * t;
     }
 
 private:
 
-    Vector3f origin;
-    Vector3f direction;
+    Vector3d origin;
+    Vector3d direction;
 
 };
